refactor: modularize PlantRun application
