+++ custom_components/plantrun/manifest.json
@@ -12,5 +12,5 @@
     "requirements": [
         "beautifulsoup4>=4.12.0"
     ],
-    "version": "0.2.0"
+    "version": "0.3.0"
 }
