+++ tests/test_panel_autocomplete.py
@@ -1,13 +1,11 @@
 import unittest
-from pathlib import Path
 
-ROOT = Path(__file__).resolve().parents[1]
-PANEL_PATH = ROOT / "custom_components" / "plantrun" / "www" / "plantrun-panel.js"
+from tests.dashboard_js_test_utils import load_panel_source
 
 
 class PanelAutocompleteTest(unittest.TestCase):
     def test_cultivar_input_supports_keyboard_click_and_stable_async_selection(self):
-        source = PANEL_PATH.read_text(encoding="utf-8")
+        source = load_panel_source()
 
         self.assertIn("_handleKeydown(event)", source)
         self.assertIn('event.key === "Enter" || event.key === "Tab"', source)
